move connection specific encoding and decoding
